clean up control methods
