+++ src/control_methods/negative_control/script.py
@@ -1,8 +1,6 @@
 import pandas as pd
 import anndata as ad
-import sys
 import numpy as np
-import sys 
 
 ## VIASH START
 par = {
@@ -11,48 +9,50 @@ par = {
   "tf_all": "resources/grn_benchmark/prior/tf_all.csv",
   "max_n_links": 50000
 }
+meta = {
+    'resources_dir':'src/control_methods/negative_control'
+}
 ## VIASH END
 
-import argparse
-parser = argparse.ArgumentParser(description="Process multiomics RNA data.")
-parser.add_argument('--rna', type=str, help='Path to the multiomics RNA file')
-parser.add_argument('--prediction', type=str, help='Path to the prediction file')
-parser.add_argument('--resources_dir', type=str, help='Path to the prediction file')
-parser.add_argument('--tf_all', type=str, help='Path to the tf_all')
-parser.add_argument('--num_workers', type=str, help='Number of cores')
-parser.add_argument('--max_n_links', type=str, help='Number of top links to retain')
-parser.add_argument('--normalize', action='store_true')
-args = parser.parse_args()
-if args.rna:
-    par['rna'] = args.rna
-if args.prediction:
-    par['prediction'] = args.prediction
-if args.tf_all:
-    par['tf_all'] = args.tf_all
-if args.num_workers:
-    par['num_workers'] = args.num_workers
-if args.max_n_links:
-    par['max_n_links'] = int(args.max_n_links)
-if args.resources_dir:
-    meta['resources_dir'] = args.resources_dir  
-print(par)
+def process_links(net, par):
+    net = net[net.source!=net.target]
+    try:
+        net = net.sample(par['max_n_links'])
+    except:
+        net = net.sample(par['max_n_links'], replace=True)
+    return net
 
-try:
-    sys.path.append(meta['resources_dir'])
-except:
-    meta = {
-    'resources_dir':'src/control_methods/negative_control'
+print('Reading input data')
+rna_all = ad.read_h5ad(par["rna_all"])
+gene_names = rna_all.var_names.to_numpy()
+tf_all = np.loadtxt(par['tf_all'], dtype=str)
+
+n_tf = 500
+tfs = tf_all[:n_tf]
+
+def create_negative_control(gene_names) -> np.ndarray:
+    ratio = [.98, .01, 0.01]
+    
+    net = np.random.choice([0, -1, 1], size=((len(gene_names), n_tf)),p=ratio)
+    net = pd.DataFrame(net, index=gene_names, columns=tfs)
+    return net
+
+print('Inferring GRN')
+net = create_negative_control(gene_names)
+pivoted_net = net.reset_index().melt(id_vars='index', var_name='source', value_name='weight')
+pivoted_net = pivoted_net.rename(columns={'index': 'target'})
+pivoted_net = pivoted_net[pivoted_net['weight'] != 0]
+pivoted_net = process_links(pivoted_net, par)
+
+print('Output GRN')
+pivoted_net['weight'] = pivoted_net['weight'].astype(str)
+output = ad.AnnData(
+    X=None,
+    uns={
+        "method_id": meta['name'],
+        "dataset_id": rna_all.uns['dataset_id'],
+        "prediction": pivoted_net[["source", "target", "weight"]]
     }
-    sys.path.append(meta['resources_dir'])
-from main import main
-
-if __name__ == '__main__':
-    dataset_id = ad.read_h5ad(par['rna'], backed='r').uns['dataset_id']
-
-    net = main(par)
-
-    print('Output GRN')
-    net['weight'] = net['weight'].astype(str)
-    output = ad.AnnData(X=None, uns={"method_id": 'negative_control', "dataset_id": dataset_id, "prediction": net[["source", "target", "weight"]]})
-    output.write(par['prediction'])
+)
+output.write_h5ad(par['prediction'])
