+++ src/control_methods/pearson_corr/config.vsh.yaml
@@ -1,6 +1,5 @@
 __merge__: ../../api/comp_control_method.yaml
 
-
 name: pearson_corr
 info:
   label: pearson_corr
